+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.1] Q3 2022
+- create location only once and save as reported property
 - ignored RUSTSEC-2020-0071
 - removed unused dependency to notify